refactor: refactor member location add sql restriction
