+++ src/main/java/com/shareit/shareit/domain/Repository/MemberRepository.java
@@ -2,7 +2,7 @@ package com.shareit.shareit.domain.Repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.shareit.shareit.domain.entity.Member;
+import com.shareit.shareit.member.domain.entity.Member;
 
 public interface MemberRepository extends JpaRepository<Member, Long> {
